requirements.txt
Flask
uwsgi
pytest
python-dateutil
